requirements.txt

# Discord Bot Dependencies
py-cord>=2.6.1
python-dotenv>=1.0.0

# Music functionality
yt-dlp>=2025.7.21
PyNaCl>=1.5.0
async-timeout>=5.0.1
ffmpeg-python>=0.2.0

# Web scraping and search
googlesearch-python>=1.3.0
beautifulsoup4>=4.13.4
requests>=2.32.4
lxml>=5.4.1

# HTTP client dependencies (required by py-cord)
aiohttp>=3.12.15
aiohappyeyeballs>=2.6.1
aiosignal>=1.4.0
attrs>=25.3.0
frozenlist>=1.7.0
multidict>=6.6.3
propcache>=0.3.2
yarl>=1.20.1

# Additional core dependencies
typing-extensions>=4.14.1
idna>=3.10
urllib3>=2.5.0
certifi>=2025.8.3
charset-normalizer>=3.4.2
soupsieve>=2.7

# Audio processing
cffi>=1.17.1
pycparser>=2.22

# Data handling and storage
asyncio-mqtt>=0.16.2
python-dateutil>=2.9.0
pytz>=2025.1

# Error handling and logging
colorama>=0.4.6
